feat: SSGされた静的HTMLにprogressive enhancementスクリプトを注入してSPAを回復
