+++ src/reroute.ts
@@ -1,10 +1,11 @@
+import { scrollToId } from './lib/render';
 import { toState } from './state';
 
-export function watchPopState(reroute: () => void): void {
+function watchPopState(reroute: () => void): void {
   window.addEventListener('popstate', () => reroute());
 }
 
-export function watchSearchForm(reroute: () => void): void {
+function watchSearchForm(reroute: () => void): void {
   const searchFormElement = document.querySelector('.el_search_form');
   const searchInputElement = <HTMLInputElement>(
     document.querySelector('.el_search_input')
@@ -23,4 +24,25 @@ export function watchSearchForm(reroute: () => void): void {
     );
     reroute();
   });
+}
+
+function watchInternalLinkClicks(reroute: () => void): void {
+  document.addEventListener('click', (e) => {
+    if (!(e.target instanceof Element)) return;
+    const a = e.target.closest<HTMLAnchorElement>(
+      'a[href^="#"], a[href^="/?"], a[href="/"], a[href^="/posts/"]'
+    );
+    if (a) {
+      e.preventDefault();
+      window.history.pushState(undefined, '', a.href);
+      reroute();
+      scrollToId(a.hash.replace('#', ''));
+    }
+  });
+}
+
+export function registerRerouter(reroute: () => void): void {
+  watchPopState(reroute);
+  watchSearchForm(reroute);
+  watchInternalLinkClicks(reroute);
 }
